feat: Add new volunteering experience
Added a new entry to the volunteering section of the portfolio. The new entry is for my role as a Student Rep at Leeds Beckett University. As a Student Rep, I represented student views on education, liaised with staff and the Students' Union, and participated in improving academic representation. My duties included understanding student experiences, discussing solutions with staff and students, and reporting outcomes.

This commit also includes an image for the new entry and updates to file paths.