Updated LazyImage component and styles
Enhanced the LazyImage component by adding a caption feature. The 'cap' prop can now be passed to display a caption under the image. If 'cap' is set to true, the alt text will be used as the caption. Also updated corresponding SCSS file to style the new caption element and made adjustments for better image display.