Updated link path in oldWebsite page
The href attribute of the Link component in the oldWebsite page has been updated. The previous path was "/fun/hobbies/website/down.html", and it's now changed to "/fun/website/down.html". This change ensures that users are directed to the correct location when they click on the sample link for the old website.