Added blank page and updated global styles
A new blank page component has been added to the application. Additionally, some changes have been made to the global styles: The maximum width of the HTML element is now set to 100vw, and a similar change has been applied to the body element. Also, an adjustment was made to the margin property of the body element.