Added LazyImage component and updated metadata
A new LazyImage component has been added to the Home function in page.tsx. This component is used to display an image with a specific aspect ratio. The source of this image is 'opengraph.webp'. In addition, the default OpenGraph object in Metadata.ts has been updated to use '/opengraph.webp' as its image instead of 'favico.png'.