Refactored font imports and updated default font
The changes in this commit involve a refactoring of the way fonts are imported into the project. The @font-face declarations for Bagnard and OfficeCodePro have been moved from globals.scss to a new file, _fonts.scss. This change helps to keep our stylesheets more organized and maintainable.

Additionally, the default font for the html element has been switched from OfficeCodePro to Bagnard. This update may affect the overall look and feel of text elements across the application.